# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ShopTik API
  version: 0.1.0
  description: API MVP cho web/PWA. Xác thực dùng phiên máy chủ trong cookie HttpOnly.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
  /csrf:
    get:
      security: []
      summary: Lấy CSRF token và thiết lập CSRF cookie
      responses:
        "200":
          description: Token cho header X-CSRF-Token
          content:
            application/json:
              schema:
                type: object
                properties:
                  csrfToken: { type: string }
                required: [csrfToken]
  /auth/register:
    post:
      security: []
      summary: Đăng ký bằng email và gửi OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "202": { description: OTP đã được gửi }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /auth/verify-email:
    post:
      security: []
      summary: Xác nhận email và tạo phiên
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                code: { type: string, pattern: "^[0-9]{6}$" }
              required: [email, code]
      responses:
        "200": { description: Email đã xác nhận }
        "400": { $ref: "#/components/responses/Error" }
  /auth/login:
    post:
      security: []
      summary: Đăng nhập email và tạo phiên
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
              required: [email, password]
      responses:
        "200":
          description: Đăng nhập thành công
          headers:
            Set-Cookie:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Error" }
  /auth/logout:
    post:
      summary: Thu hồi phiên hiện tại
      responses:
        "204": { description: Đã đăng xuất }
  /auth/forgot-password:
    post:
      security: []
      summary: Yêu cầu OTP đặt lại mật khẩu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
              required: [email]
      responses:
        "202": { description: Luôn trả thông báo trung tính }
  /auth/reset-password:
    post:
      security: []
      summary: Đổi mật khẩu bằng OTP email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                code: { type: string, pattern: "^[0-9]{6}$" }
                password: { type: string, minLength: 10, maxLength: 128 }
              required: [email, code, password]
      responses:
        "204": { description: Đã đổi mật khẩu và thu hồi mọi phiên cũ }
        "400": { $ref: "#/components/responses/Error" }
  /me:
    get:
      summary: Hồ sơ và số dư tổng hợp
      responses:
        "200":
          description: Người dùng hiện tại
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  balances: { $ref: "#/components/schemas/Balances" }
        "401": { $ref: "#/components/responses/Error" }
  /products/preview:
    post:
      summary: Tra cứu sản phẩm (tên, ảnh, giá gốc, tiền hoàn dự kiến)
      description: Không ghi dữ liệu. Trả previewId dùng cho bước /products/purchase, hiệu lực 15 phút.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productUrl: { type: string, format: uri, maxLength: 2048 }
                platform: { type: string, enum: [SHOPEE, TIKTOK, LAZADA] }
              required: [productUrl]
      responses:
        "200":
          description: Thông tin sản phẩm và tiền hoàn dự kiến
          content:
            application/json:
              schema:
                type: object
                properties:
                  product: { type: object }
                  previewId: { type: string }
        "400": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /products/purchase:
    post:
      summary: Bấm Mua ngay — tạo link Affiliate từ kết quả tra cứu
      description: Đổi previewId lấy buyUrl (/go/{clickId}); lúc này mới ghi bản ghi affiliate_links.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                previewId: { type: string }
              required: [previewId]
      responses:
        "201":
          description: Link mua đã gắn mã Affiliate
          content:
            application/json:
              schema:
                type: object
                properties:
                  buyUrl: { type: string, format: uri }
                  clickId: { type: string }
                  platform: { type: string, enum: [SHOPEE, TIKTOK, LAZADA] }
        "400": { $ref: "#/components/responses/Error" }
        "410": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /me/orders:
    get:
      summary: Danh sách đơn của người dùng
      responses:
        "200": { description: Tối đa 100 đơn gần nhất }
  /me/wallet:
    get:
      summary: Số dư và lịch sử bút toán
      responses:
        "200": { description: Sổ biến động của người dùng }
  /me/withdrawals:
    get:
      summary: Lịch sử rút tiền
      responses:
        "200": { description: Tối đa 100 yêu cầu gần nhất }
  /support/missing-order:
    post:
      summary: Tạo ticket báo thiếu đơn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId: { type: string, minLength: 3, maxLength: 100 }
                description: { type: string, minLength: 20, maxLength: 3000 }
              required: [orderId, description]
      responses:
        "201": { description: Ticket đã tạo }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: aff_session
  responses:
    Error:
      description: Lỗi nghiệp vụ hoặc xác thực
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    RegisterRequest:
      type: object
      properties:
        fullName: { type: string, minLength: 2, maxLength: 100 }
        email: { type: string, format: email }
        password: { type: string, minLength: 10, maxLength: 128 }
        passwordConfirm: { type: string }
        referralCode: { type: string, maxLength: 30 }
        acceptPolicies: { type: boolean, const: true }
      required:
        [fullName, email, password, passwordConfirm, acceptPolicies]
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        fullName: { type: string }
        role: { type: string }
        status: { type: string }
        referralCode: { type: string }
      required: [id, email, fullName, role, status, referralCode]
    Balances:
      type: object
      properties:
        pending: { type: integer, minimum: 0 }
        available: { type: integer, minimum: 0 }
        held: { type: integer, minimum: 0 }
        paid: { type: integer, minimum: 0 }
      required: [pending, available, held, paid]
    AffiliateLink:
      type: object
      properties:
        clickId: { type: string }
        buyUrl: { type: string, format: uri }
        normalizedUrl: { type: string, format: uri }
        subId: { type: string }
      required: [clickId, buyUrl, normalizedUrl, subId]
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
          required: [code, message, requestId]
      required: [error]